Set rate type to double percision

+++ pg/psql_dump.sql
@@ -1,6 +1,6 @@
 CREATE TABLE rates (
   id serial PRIMARY KEY,
   protocol VARCHAR(40) NOT NULL,
-  rate BIGINT NOT NULL,
+  rate DOUBLE PRECISION NOT NULL,
   ts TIMESTAMP
 )
